+++ Sources/Parsing/MappingEntry.swift
@@ -1,9 +1,0 @@
-extension PureYAML.Parsing {
-    struct MappingEntry: Equatable {
-        var key: String
-        var keyStyle: ScalarStyle
-        var keyOffset: Int
-        var value: String
-        var valueOffset: Int
-    }
-}