+++ Sources/Parsing/ScalarStyle.swift
@@ -1,5 +1,7 @@
 extension PureYAML.Parsing {
     enum ScalarStyle: String, Equatable, CustomStringConvertible {
+        case folded
+        case literal
         case plain
         case singleQuoted
         case doubleQuoted